update prefix and postfixconvertor.html
The current version pops from the stack only when the element has lower precedence that the operator in the top of the stack.. 
But actually, the popping should be done even when the element in the top of the stack has the same priority (except for '^' operator as it has same priority. 
References: 
GFG: https://www.geeksforgeeks.org/convert-infix-expression-to-postfix-expression/
Quescol: https://quescol.com/data-structure/infix-to-postfix 
This issue has been fixed in the ds.js file and the same algorithmic change has been documented in the prefixAndPostfixConvertor.html file.